# web/static/api/openapi.yml
openapi: 3.0.0
info:
  title: AzuraCast
  description: 'AzuraCast is a standalone, turnkey web radio management tool. Radio stations hosted by AzuraCast expose a public API for viewing now playing data, making requests and more.'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: '0.11'
servers:
  -
    url: 'https://demo.azuracast.com/api'
    description: 'AzuraCast Public Demo Server'
paths:
...
  /admin/roles:
    get:
      tags:
        - 'Administration: Roles'
      description: 'List all current roles in the system.'
      operationId: 'App\Controller\Api\Admin\RolesController::fromArray'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    post:
      tags:
        - 'Administration: Roles'
      description: 'Create a new role.'
      operationId: 'App\Controller\Api\Admin\RolesController::fromArray'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
  '/admin/role/{id}':
    get:
      tags:
        - 'Administration: Roles'
      description: 'Retrieve details for a single current role.'
      operationId: 'App\Controller\Api\Admin\RolesController::fromArray'
      parameters:
        -
          name: id
          in: path
          description: 'Role ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    put:
      tags:
        - 'Administration: Roles'
      description: 'Update details of a single role.'
      operationId: 'App\Controller\Api\Admin\RolesController::fromArray'
      parameters:
        -
          name: id
          in: path
          description: 'Role ID'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Status'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    delete:
      tags:
        - 'Administration: Roles'
      description: 'Delete a single role.'
      operationId: 'App\Controller\Api\Admin\RolesController::fromArray'
      parameters:
        -
          name: id
          in: path
          description: 'Role ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Status'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
  /admin/settings:
    get:
      tags:
        - 'Administration: Settings'
      description: 'List the current values of all editable system settings.'
      operationId: 'App\Controller\Api\Admin\SettingsController::listAction'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Admin_Settings'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    put:
      tags:
        - 'Administration: Settings'
      description: 'Update settings to modify any settings provided.'
      operationId: 'App\Controller\Api\Admin\SettingsController::updateAction'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Api_Admin_Settings'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Status'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
...
  /admin/storage_locations:
    get:
      tags:
        - 'Administration: Storage Locations'
      description: 'List all current storage locations in the system.'
      operationId: 'App\Controller\Api\Admin\StorageLocationsController::listAction'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Api_Admin_StorageLocation'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    post:
      tags:
        - 'Administration: Storage Locations'
      description: 'Create a new storage location.'
      operationId: 'App\Controller\Api\Admin\StorageLocationsController::listAction'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Api_Admin_StorageLocation'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Admin_StorageLocation'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
  '/admin/storage_location/{id}':
    get:
      tags:
        - 'Administration: Storage Locations'
      description: 'Retrieve details for a single storage location.'
      operationId: 'App\Controller\Api\Admin\StorageLocationsController::listAction'
      parameters:
        -
          name: id
          in: path
          description: 'User ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Admin_StorageLocation'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    put:
      tags:
        - 'Administration: Storage Locations'
      description: 'Update details of a single storage location.'
      operationId: 'App\Controller\Api\Admin\StorageLocationsController::listAction'
      parameters:
        -
          name: id
          in: path
          description: 'Storage Location ID'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Api_Admin_StorageLocation'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Status'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    delete:
      tags:
        - 'Administration: Storage Locations'
      description: 'Delete a single storage location.'
      operationId: 'App\Controller\Api\Admin\StorageLocationsController::listAction'
      parameters:
        -
          name: id
          in: path
          description: 'Storage Location ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Status'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
...
  '/station/{station_id}/queue':
    get:
      tags:
        - 'Stations: Queue'
      description: 'Return information about the upcoming song playback queue.'
      operationId: 'App\Controller\Api\Stations\QueueController::listAction'
      parameters:
        -
          $ref: '#/components/parameters/station_id_required'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Api_StationQueueDetailed'
        '404':
          description: 'Station not found'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
  '/station/{station_id}/queue/{id}':
    get:
      tags:
        - 'Stations: Queue'
      description: 'Retrieve details of a single queued item.'
      parameters:
        -
          name: id
          in: path
          description: 'Queue Item ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_StationQueueDetailed'
        '404':
          description: 'Station or Queue ID not found'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
    delete:
      tags:
        - 'Stations: Queue'
      description: 'Delete a single queued item.'
      parameters:
        -
          name: id
          in: path
          description: 'Queue Item ID'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Status'
        '404':
          description: 'Station or Queue ID not found'
        '403':
          description: 'Access denied'
      security:
        -
          api_key: []
...
  '/station/{station_id}/requests':
    get:
      tags:
        - 'Stations: Song Requests'
      description: 'Return a list of requestable songs.'
      operationId: 'App\Controller\Api\Stations\RequestsController::listAction'
      parameters:
        -
          $ref: '#/components/parameters/station_id_required'
      responses:
        '200':
          description: Success
        '404':
          description: 'Station not found'
        '403':
          description: 'Station does not support requests'
...
components:
  schemas:
...
    Api_Admin_Settings:
      properties:
        base_url:
          description: 'Site Base URL'
          type: string
        instance_name:
          description: 'AzuraCast Instance Name'
          type: string
          nullable: true
        timezone:
          description: 'System Default Time Zone'
          type: string
          nullable: true
        prefer_browser_url:
          description: 'Prefer Browser URL (If Available)'
          type: integer
        use_radio_proxy:
          description: 'Use Web Proxy for Radio'
          type: integer
        history_keep_days:
          description: 'Days of Playback History to Keep'
          type: integer
        always_use_ssl:
          description: 'Always Use HTTPS'
          type: integer
        api_access_control:
          description: 'API "Access-Control-Allow-Origin" header'
          type: string
        analytics:
          description: 'Listener Analytics Collection'
          type: string
        central_updates_channel:
          description: 'Check for Updates and Announcements'
          type: integer
        public_theme:
          description: 'Base Theme for Public Pages'
          type: string
        hide_album_art:
          description: 'Hide Album Art on Public Pages'
          type: integer
        homepage_redirect_url:
          description: 'Homepage Redirect URL'
          type: string
          nullable: true
        default_album_art_url:
          description: 'Default Album Art URL'
          type: string
          nullable: true
        hide_product_name:
          description: 'Hide AzuraCast Branding on Public Pages'
          type: integer
        custom_css_public:
          description: 'Custom CSS for Public Pages'
          type: string
          nullable: true
        custom_js_public:
          description: 'Custom JS for Public Pages'
          type: string
          nullable: true
        custom_css_internal:
          description: 'Custom CSS for Internal Pages'
          type: string
          nullable: true
      type: object
    Api_Admin_StorageLocation:
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          example: station_media
        adapter:
          type: string
          example: local
        path:
          type: string
          example: /var/azuracast/stations/azuratest_radio/media
          nullable: true
        s3CredentialKey:
          type: string
          example: your-key-here
          nullable: true
        s3CredentialSecret:
          type: string
          example: your-secret-here
          nullable: true
        s3Region:
          type: string
          example: your-region
          nullable: true
        s3Version:
          type: string
          example: latest
          nullable: true
        s3Bucket:
          type: string
          example: your-bucket-name
          nullable: true
        s3Endpoint:
          type: string
          example: 'https://your-region.digitaloceanspaces.com'
          nullable: true
        storageQuota:
          type: string
          example: '50 GB'
          nullable: true
        storageUsed:
          type: string
          example: '1 GB'
          nullable: true
        uri:
          description: 'The URI associated with the storage location.'
          type: string
          example: /var/azuracast/www
        stations:
          description: 'The stations using this storage location, if any.'
          type: array
          items:
            type: string
            example: 'AzuraTest Radio'
          nullable: true
      type: object
...
    Api_Song:
      properties:
        id:
          description: 'The song''s 32-character unique identifier hash'
          type: string
          example: 9f33bbc912c19603e51be8e0987d076b
        text:
          description: 'The song title, usually "Artist - Title"'
          type: string
          example: 'Chet Porter - Aluko River'
        artist:
          description: 'The song artist.'
          type: string
          example: 'Chet Porter'
        title:
          description: 'The song title.'
          type: string
          example: 'Aluko River'
        album:
          description: 'The song album.'
          type: string
          example: 'Moving Castle'
        genre:
          description: 'The song genre.'
          type: string
          example: Rock
        lyrics:
          description: 'Lyrics to the song.'
          type: string
          example: ''
        art:
          description: 'URL to the album artwork (if available).'
          example: 'https://picsum.photos/1200/1200'
        custom_fields:
          type: array
          items:
            type: string
            example: custom_field_value
      type: object
...
    Api_StationQueue:
      properties:
        cued_at:
          description: 'UNIX timestamp when playback is expected to start.'
          type: integer
          example: 1606584191
        duration:
          description: 'Duration of the song in seconds'
          type: integer
          example: 180
        playlist:
          description: 'Indicates the playlist that the song was played from, if available, or empty string if not.'
          type: string
          example: 'Top 100'
          nullable: true
        is_request:
          description: 'Indicates whether the song is a listener request.'
          type: boolean
        song:
          $ref: '#/components/schemas/Api_Song'
      type: object
    Api_StationQueueDetailed:
      allOf:
        -
          $ref: '#/components/schemas/Api_StationQueue'
        -
          properties:
            links:
              type: array
              items:
                type: string
                example: 'http://localhost/api/stations/1/queue/1'
            autodj_custom_uri:
              description: 'Custom AutoDJ playback URI, if it exists.'
              type: string
              example: ''
              nullable: true
          type: object
...
    Api_Status:
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Changes saved successfully.'
        formatted_message:
          type: string
          example: '<b>Changes saved successfully.</b>'
      type: object
...
    Role:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Super Administrator'
        permissions:
          items: {  }
      type: object
...
  parameters:
    station_id_required:
      name: station_id
      in: path
      description: 'The station ID'
      required: true
      schema:
        type: integer
        format: int64
      example: 1
  securitySchemes:
    api_key:
      type: apiKey
      name: X-API-Key
      in: header
...